Gemmi compatibility

+++ pandda_gemmi/cnn/event.py
@@ -149,10 +149,10 @@ class EventScorer:
 
         if self.debug is True:
             centroid = event.centroid
-            z_mean = round(np.mean(zmap_sample), 2)
-            z_std = round(np.std(zmap_sample), 2)
-            x_mean = round(np.mean(zmap_sample), 2)
-            x_std = round(np.std(zmap_sample), 2)
+            z_mean = round(float(np.mean(zmap_sample), 2))
+            z_std = round(float(np.std(zmap_sample), 2))
+            x_mean = round(float(np.mean(zmap_sample), 2))
+            x_std = round(float(np.std(zmap_sample), 2))
             m_num = np.sum(density_mask)
             print(f'\tEvent {round(centroid[0], 2)} {round(centroid[1], 2)} {round(centroid[2], 2)}: z: {round(z_mean, 2)} {round(z_std, 2)}: x: {round(x_mean, 2)} {round(x_std, 2)}; mask {m_num}')
         # Run the model
